Fixed Load Order
For some reason, the load order was full of bollocks, It loaded Clients first then procceded to load routes? what about the Distribution centers?
In other words, it was assigning clients to non-existent distribution centers. this and with the fixes to the Graph as it was not adding distribution centers correctly solves the issue.