Added untested Backtracking Algorithm + Heap Queue
Pretty much made the basis of the Backtracking Algorithm, however i am not sure if it works or not, specially the part in which it determines if it is a Leaf or Not and what it does after, however the rest should be fine. Also added the Min-Heap Priority Queue that is usually used in Backtracking Algorithms with Trimming.